requirements.txt
fastapi
uvicorn
pandas
google-genai